Restructured hierarchy of the folders
