+++ ToDoList/ToDoList/Modules/TaskManager/Builder/TaskManagerBuilder.swift
@@ -1,0 +1,22 @@
+import UIKit
+
+final class TaskManagerBuilder {
+
+    static let shared = TaskManagerBuilder()
+
+    private init() {}
+
+    func buildTaskManager() -> TaskManagerViewControllerImp? {
+        let storage = StorageImp()
+        let taskManagerVC = TaskManagerViewControllerImp(
+            nibName: String(describing: TaskManagerViewControllerImp.self),
+            bundle: nil
+        )
+        let taskManagerPresenter = TaskManagerPresenterImp(
+            viewController: taskManagerVC,
+            storage: storage
+        )
+        taskManagerVC.presenter = taskManagerPresenter
+        return taskManagerVC
+    }
+}
