+++ ToDoList/ToDoList/Modules/TaskManager/Core/Controllers/Protocols/ViewControllerProtocol.swift
@@ -1,0 +1,8 @@
+//
+//  ViewControllerProtocol.swift
+//  ToDoList
+//
+//  Created by Никита Данилович on 03.05.2023.
+//
+
+import Foundation
