+++ ToDoList/ToDoList/Modules/TaskManager/Core/Controllers/TaskTableViewCell/TaskTableViewCell.xib
