+++ ToDoList/ToDoList/Modules/TaskManager/Core/Presenter/Protocols/PresenterProtocol.swift
@@ -1,0 +1,8 @@
+//
+//  PresenterProtocol.swift
+//  ToDoList
+//
+//  Created by Никита Данилович on 03.05.2023.
+//
+
+import Foundation
